backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
Pillow>=9.0.0
imagehash>=4.0.0
pydantic>=2.0.0
aiofiles>=0.8.0